+++ RELEASES.md
@@ -1,3 +1,9 @@
+Version 0.5.3 (2026-01-25)
+==========================
+* Added from OwnedFd and from RawFd and from Self to OwnedFd. Functions which performs conversion from OwnedFd to Self perform check of the socket type.
+* `NonblockingUnixSeqpacketListener` and `NonblockingUnixSeqpacketConn` is now have inside an instance of `UnixSeqpacketConn` and `UnixSeqpacketListener` respectivly which is set to non-blocking.
+* Reorganised some functions
+
 Version 0.5.2 (2026-01-21)
 ==========================
 * Fixed compilation errors on OpenBSD